+++ src/gradle/kotlin/no/f12/codenavigator/gradle/CodeNavigatorPlugin.kt
@@ -101,6 +101,9 @@ class CodeNavigatorPlugin : Plugin<Project> {
                 if (this is RenamePropertyTask) {
                     openRewriteClasspath.from(openRewriteConfig)
                 }
+                if (this is SafeDeleteTask) {
+                    psiClasspath.from(psiConfig)
+                }
             }
         }
 
@@ -149,6 +152,7 @@ class CodeNavigatorPlugin : Plugin<Project> {
             "rename-method" to RenameMethodTask::class.java,
             "move-class" to MoveClassTask::class.java,
             "rename-property" to RenamePropertyTask::class.java,
+            "safe-delete" to SafeDeleteTask::class.java,
             "help" to CodeNavigatorHelpTask::class.java,
             "agent-help" to AgentHelpTask::class.java,
             "config-help" to ConfigHelpTask::class.java,